Added mod log admin panel.

+++ src/bot/slashCommands/moderation/ban.ts
@@ -51,45 +51,48 @@ export default class Ban extends SlashCommand {
         const reason = interaction.options.getString("reason") ?? "No reason provided.";
         const messages = new Collection<Snowflake, Message<boolean>>();
         const channel = interaction.channel as GuildTextBasedChannel;
+		const purge = interaction.options.getBoolean("purge") ?? false;
 
-        for (const message of (
-			await channel.messages.fetch({
-				limit: 100,
-			} as ChannelLogsQueryOptions)
-		).values()) {
-			let completed = [];
-			if (interaction.options.getUser("user"))
-				completed.push(interaction.options.getUser("user")!.id === message.author.id);
-			if (completed.every((c) => c)) messages.set(message.id, message);
-		}
-		let deleted = new Collection<Snowflake, Message>();
-		if (messages.size === 0)
-			return interaction.reply(
-				this.client.functions.generateErrorMessage({
-					title: "No Messages Match Criteria",
-					description: "There are no messages that match the criteria you provided!"
-				})
-			);
-		else if (messages.size === 1) {
-			const d = await messages.first()!.delete();
-			deleted.set(d.id, d);
-		} else deleted = await channel.bulkDelete(messages, true);
-		this.client.emit(
-			"purge",
-			interaction.guild,
-			interaction.member,
-			messages,
-			interaction.options.getString("reason")
-		);
-		let users = new Collection<Snowflake, User>();
-		let purgedUserCount = new Collection<Snowflake, number>();
-		deleted.forEach((message) => {
-			if (!users.get(message.author.id)) {
-				users.set(message.author.id, message.author);
-				purgedUserCount.set(message.author.id, 0);
+		if (purge) {
+			for (const message of (
+				await channel.messages.fetch({
+					limit: 100,
+				} as ChannelLogsQueryOptions)
+			).values()) {
+				let completed = [];
+				if (interaction.options.getUser("user"))
+					completed.push(interaction.options.getUser("user")!.id === message.author.id);
+				if (completed.every((c) => c)) messages.set(message.id, message);
 			}
-			purgedUserCount.set(message.author.id, purgedUserCount.get(message.author.id)! + 1);
-		});
+			let deleted = new Collection<Snowflake, Message>();
+			if (messages.size === 0)
+				return interaction.reply(
+					this.client.functions.generateErrorMessage({
+						title: "No Messages Match Criteria",
+						description: "There are no messages that match the criteria you provided!"
+					})
+				);
+			else if (messages.size === 1) {
+				const d = await messages.first()!.delete();
+				deleted.set(d.id, d);
+			} else deleted = await channel.bulkDelete(messages, true);
+			this.client.emit(
+				"purge",
+				interaction.guild,
+				interaction.member,
+				messages,
+				interaction.options.getString("reason")
+			);
+			let users = new Collection<Snowflake, User>();
+			let purgedUserCount = new Collection<Snowflake, number>();
+			deleted.forEach((message) => {
+				if (!users.get(message.author.id)) {
+					users.set(message.author.id, message.author);
+					purgedUserCount.set(message.author.id, 0);
+				}
+				purgedUserCount.set(message.author.id, purgedUserCount.get(message.author.id)! + 1);
+			});
+		}
         
         
         //await user.ban({ reason });
